% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_coefficient.R
\name{binomial.coefficient.factorial}
\alias{binomial.coefficient.factorial}
\title{Calculates the binomial coefficient using the factorial method.}
\usage{
binomial.coefficient.factorial(n, k)
}
\arguments{
\item{n}{Number of possibilities}

\item{k}{number of unordered outcomes}
}
\value{
The binomial coefficient
}
\description{
The binomial coefficient equation (in compact form) is defined as:
\deqn{\binom{n}{k} = \frac{n!}{k!(n-k)!} \qquad 0 \leq k \leq n}
}
\references{
Binomial coefficient. (2017, April 17). In Wikipedia, The Free
  Encyclopedia. From
  https://en.wikipedia.org/w/index.php?title=Binomial_coefficient&oldid=775905810
   Binomial coefficients. Encyclopedia of Mathematics. From
  http://www.encyclopediaofmath.org/index.php?title=Binomial_coefficients&oldid=39155
   Weisstein, Eric W. "Binomial Coefficient." From MathWorld--A Wolfram Web
  Resource. http://mathworld.wolfram.com/BinomialCoefficient.html
}
