% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_coefficient.R
\name{binomial.coefficient}
\alias{binomial.coefficient}
\title{Calculates the binomial coefficient. Wraps all three binomial coefficient
functions and defaults to the multiplicative method. Other methods can be
specified by the method argument.}
\usage{
binomial.coefficient(n, k, method = c("multiplicative", "recursive",
  "factorial"))
}
\arguments{
\item{n}{Number of possibilities}

\item{k}{number of unordered outcomes}

\item{method}{Sets method to calculate binomial coefficient. Defaults to
multiplicative. Be warned the recursive method for larger integers can take
an exorbitant amount of time to complete.}
}
\value{
The binomial coefficient
}
\description{
The binomial coefficient \eqn{\binom{n}{k}} is defined as the number of ways
of selecting a set of \eqn{k} elements (unordered) from a set of \eqn{n}
elements and is read as \eqn{n} choose \eqn{k}. The compact form of the
binomial coefficient equation is defined as: \deqn{\binom{n}{k} =
\frac{n!}{k!(n - k)!} \qquad 0 \leq k \leq n}
}
\examples{
binomial.coefficient(4, 2)
binomial.coefficient(10, 5)
binomial.coefficient(10, 5, 'multiplicative')
binomial.coefficient(10, 5, 'factorial')
}
\references{
Press, W., Teukolsky, S., Vetterling, W., & Flannery, B. (2007).
  Numerical recipes (3rd ed.). Cambridge: Cambridge University Press.
  Weisstein, Eric W. "Binomial Coefficient." From MathWorld--A Wolfram Web
  Resource.http://mathworld.wolfram.com/BinomialCoefficient.html
}
