% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{binomial_recursive}
\alias{binomial_recursive}
\title{Calculates the binomial coefficient using a recursive method.}
\usage{
binomial_recursive(n, k)
}
\arguments{
\item{n}{Number of possibilities}

\item{k}{number of unordered outcomes}
}
\value{
The binomial coefficient
}
\description{
The recursive method of the binomial coefficient calculation is defined as:
\deqn{\binom{n}{k} = \binom{n - 1}{n - k} + \binom{n - 1}{k} \qquad for n, k:
1 \leq k \leq n - 1} With boundary values: \deqn{\binom{n}{0} = \binom{n}{n}
= 1}
}
\references{
Binomial coefficient. (2017, April 17). In Wikipedia, The Free
  Encyclopedia. From
  https://en.wikipedia.org/w/index.php?title=Binomial_coefficient&oldid=775905810
   Binomial coefficients. Encyclopedia of Mathematics. From
  http://www.encyclopediaofmath.org/index.php?title=Binomial_coefficients&oldid=39155
   Weisstein, Eric W. "Binomial Coefficient." From MathWorld--A Wolfram Web
  Resource. http://mathworld.wolfram.com/BinomialCoefficient.html
}
