% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{catalans_constant}
\alias{catalans_constant}
\title{Approximates Catalan's Constant, K.}
\usage{
catalans_constant(k = 10L)
}
\arguments{
\item{k}{number of summations to perform, defaults to 10.}
}
\value{
Approximated Catalan's Constant
}
\description{
Approximates Catalan's Constant, K.
}
