% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cullen}
\alias{cullen}
\title{Returns the Cullen number integer sequence up to a given value of n.}
\usage{
cullen(n)
}
\arguments{
\item{n}{Specifies the length of the Cullen number sequence to return}
}
\value{
vector of length n
}
\description{
Cullen numbers are a special case of Proth numbers that have the form:
\deqn{C_n = 2^n n + 1} The first few Cullen numbers are 3, 9, 25, 65, 161, ...
}
\references{
Cullen number. (2018, January 22). In Wikipedia, The Free
  Encyclopedia. Retrieved 14:17, January 27, 2018, from
  https://en.wikipedia.org/w/index.php?title=Cullen_number&oldid=821774732
  Weisstein, Eric W. "Cullen Number." From MathWorld--A Wolfram Web
  Resource. http://mathworld.wolfram.com/CullenNumber.html
}
