% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor.R
\name{factor}
\alias{factor}
\title{Performs integer factorization.}
\usage{
factor(n, method = c("trial", "fermat", "pollardrho"))
}
\arguments{
\item{n}{Integer to be factored into product of smaller integers.}

\item{method}{Specifies the method to use when performing integer
factorization. Available methods are trial division, Fermat
factorization and Pollard-Rho factorization.}
}
\value{
The factors of the integer, should they exist
}
\description{
Performs integer factorization.
}
\examples{
factor(4)
factor(9, 'trial')
factor(36, 'pollardrho')
}
