% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorial.R
\name{factorial}
\alias{factorial}
\title{Function for calculating factorials using the standard approach as explained
below.}
\usage{
factorial(n)
}
\arguments{
\item{n}{Desired integer to compute factorial}
}
\description{
Factorials are denoted for a positive integer \eqn{x} as \eqn{x!} and are
defined as: \deqn{x! = (x)(x - 1)(x - 2) \cdots (2)(1)} For example, the
factorial of 5 is written as: 5! = (5)(4)(3)(2)(1) = 120
}
\examples{
factorial(10)
factorial(50)
}
\references{
Press, W., Teukolsky, S., Vetterling, W., & Flannery, B. (2007).
  Numerical recipes (3rd ed.). Cambridge: Cambridge University Press.
  Weisstein, Eric W. "Factorial." From MathWorld--A Wolfram Web Resource.
  http:#mathworld.wolfram.com/Factorial.html
}
