% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorial.R
\name{falling.factorial}
\alias{falling.factorial}
\title{Computes the falling factorial}
\usage{
falling.factorial(x, n)
}
\arguments{
\item{x}{Integer. The value will be rounded down if the value is not an
integer.}

\item{n}{Integer. The value will be rounded down if the value is not an
integer.}
}
\value{
If x is an integer, the falling factorial will be returned. If x is
  a string, the symbolic representation of the falling factorial is returned.
}
\description{
The falling factorial, denoted as \eqn{(x)_{n}} (or \eqn{x^{\underline{n}}})
is defined as the following: \deqn{(x)_n = x(x - 1) \cdots (x - (n - 1))} The
first few falling factorials are then: \deqn{(x)_0 = 1} \deqn{(x)_1 = x}
\deqn{(x)_2 = x(x - 1)} \deqn{(x)_3 = x(x - 1)(x - 2)} \deqn{(x)_4 = x(x -
1)(x - 2)(x - 3)(x - 4)}
}
\examples{
falling.factorial(10, 5)
falling.factorial(10, 2)
falling.factorial('x', 2)
falling.factorial('a', 4)
}
\references{
Falling and rising factorials. (2017, June 8). In Wikipedia, The
  Free Encyclopedia. From
  https://en.wikipedia.org/w/index.php?title=Falling_and_rising_factorials&oldid=784512036


  Weisstein, Eric W. "Falling Factorial." From MathWorld--A Wolfram Web
  Resource. http://mathworld.wolfram.com/FallingFactorial.html
}
