% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorial.R
\name{fallingfactorial}
\alias{fallingfactorial}
\title{Computes the falling factorial. If the parameter x is a character, the
function form of the falling factorial is output. If x is an integer, the
calculated falling factorial is returned.}
\usage{
fallingfactorial(x, n)
}
\arguments{
\item{x}{integer or character}

\item{n}{integer}
}
\description{
The falling factorial, denoted as \eqn{(x)_{n}} (or \eqn{x^{\underline{n}}})
is defined as the following: \deqn{(x)_n = x(x - 1) \cdots (x - (n - 1))} The
first few falling factorials are then: \deqn{(x)_0 = 1} \deqn{(x)_1 = x}
\deqn{(x)_2 = x(x - 1)} \deqn{(x)_3 = x(x - 1)(x - 2)} \deqn{(x)_4 = x(x -
1)(x - 2)(x - 3)}
}
\examples{
fallingfactorial(10, 5)
fallingfactorial(10, 6)
fallingfactorial(5, 3)
fallingfactorial("x", 4)
fallingfactorial("a", 5)
}
\references{
Falling and rising factorials. (2017, June 8). In Wikipedia, The
  Free Encyclopedia. From
  https:#en.wikipedia.org/w/index.php?title=Falling_and_rising_factorials&oldid=784512036
   Weisstein, Eric W. "Falling Factorial." From MathWorld--A Wolfram Web
  Resource. http:#mathworld.wolfram.com/FallingFactorial.html
}
