% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor.R
\name{fermat_factor}
\alias{fermat_factor}
\title{Computes the factorization of an integer \eqn{n} by Fermat's factorization
method.}
\usage{
fermat_factor(n)
}
\arguments{
\item{n}{Integer to be factored into product of smaller integers.}
}
\description{
Fermat's factorization theorem redefines a composite number \eqn{n} as the
difference of squares: \deqn{n = a^2 - b^2} Which can also be written as:
\deqn{n = (a + b)(a - b)}
}
\examples{
fermat(9)
fermat(13742)
fermat(17)
}
\references{
Barnes, C. (2004). Integer Factorization Algorithms (1st ed.).
  Corvallis, OR: Department of Physics, Oregon State University. Fermat's
  factorization method. (2017, January 31). In Wikipedia, The Free
  Encyclopedia. From
  https:#en.wikipedia.org/w/index.php?title=Fermat%27s_factorization_method&oldid=763010603
}
