% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fibonacci.R
\name{fibonacci}
\alias{fibonacci}
\title{Computes the Fibonacci sequence up to given value of n.}
\usage{
fibonacci(n, output = c("last", "array"))
}
\arguments{
\item{n}{Integer designating the stopping point of the Fibonacci sequence}

\item{output}{If 'last', the last integer in the Fibonacci sequence up to n
is returned, if 'array', the entire sequence is returned. Defaults to
'last'.}
}
\value{
Last integer in Fibonacci sequence up to \eqn{n} or the an array of
  the Fibonacci sequence up to \eqn{n}.
}
\description{
The Fibonacci sequence is defined by a recurrence relation where \eqn{f_n}
denotes the series up to :math:`n` points. \deqn{f_n = f_{n-1} + f_{n-2}}
With initial coniditions: \deqn{f_1 = 1, \qquad f_2 = 2}
}
\examples{
fibonacci(10)
fibonacci(10, 'array')
}
\references{
Moler, C. (2011). Numerical computing with MATLAB (1st ed.).
Philadelphia, Pa: Society for Industrial & Applied Mathematics.
}
