% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcd.R
\name{gcd.division}
\alias{gcd.division}
\title{Calculates the Greatest Common Divisor using the Euclidean division
algorithm.}
\usage{
gcd.division(a, b)
}
\arguments{
\item{a}{First integer}

\item{b}{Second integer}
}
\value{
The greatest common divisor
}
\description{
The Euclidean division implementation of the greatest common divisor
algorithm computes a quotient \eqn{q_k} and a remainder \eqn{r_k} at each
step \eqn{k} from the two numbers \eqn{a} and \eqn{b}. The quotient component
of Euclidean division is not used in the algorithm, thus only modulo
operations are required as that only returns the remainder.
}
\examples{
gcd.division(24, 30)
gcd.division(9, 5)
}
\references{
Cormen, T., Leiserson, C., Rivest, R., & Stein, C. (2009).
  Introduction to algorithms (3rd ed., pp. 928-930). Cambridge (Inglaterra):
  Mit Press. Euclidean algorithm. (2017, May 18). In Wikipedia, The Free
  Encyclopedia. From
  https://en.wikipedia.org/w/index.php?title=Euclidean_algorithm&oldid=780973502
   Euclidean division. (2017, May 10). In Wikipedia, The Free Encyclopedia.
  From
  https://en.wikipedia.org/w/index.php?title=Euclidean_division&oldid=779699188
}
