% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gcd}
\alias{gcd}
\title{Computes the greatest common divisor with several algorithms available. The
default algorithm is the recursive method, which is typically faster than
the others.}
\usage{
gcd(a, b, method = "recursive")
}
\arguments{
\item{a}{First integer}

\item{b}{Second integer}

\item{method}{Specifies the algorithm used to calculate the greatest common
divisior. Defaults to 'recursive', which is generally faster than other
methods. Other algorithms available include 'division' and 'subtraction'.}
}
\value{
The greatest common divisor
}
\description{
Computes the greatest common divisor with several algorithms available. The
default algorithm is the recursive method, which is typically faster than
the others.
}
