% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcd.R
\name{gcd.recursive}
\alias{gcd.recursive}
\title{Calculates the Greatest Common Divisor of two integers using the recursive
Euclidean algorithm}
\usage{
gcd.recursive(a, b)
}
\arguments{
\item{a}{First integer}

\item{b}{Second integer}
}
\value{
The greatest common divisor
}
\description{
The recursive implementation of the greatest common divisor algorithm was
described in the Elements of Euclid but may even predate that.
}
\references{
Cormen, T., Leiserson, C., Rivest, R., & Stein, C. (2009).
  Introduction to algorithms (3rd ed., pp. 928-930, 934-935). Cambridge
  (Inglaterra): Mit Press. Euclidean algorithm. (2017, May 18). In Wikipedia,
  The Free Encyclopedia. From
  https://en.wikipedia.org/w/index.php?title=Euclidean_algorithm&oldid=780973502
}
