% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gcd_extended}
\alias{gcd_extended}
\title{Implementation of the extended form of the Euclidean algorithm which computes
the greatest common divisor \eqn{d} and integers \eqn{x} and \eqn{y} such
that \eqn{ax + by = d}.}
\usage{
gcd_extended(a, b)
}
\arguments{
\item{a}{First integer}

\item{b}{Second integer}
}
\value{
vector containing coefficients (d, x, y)
}
\description{
The extended Euclidean algorithm computes the greatest common divisor,
\eqn{d} of two integers \eqn{a} and \eqn{b} as well as the coefficients
\eqn{x} and \eqn{y} such that: \deqn{d = gcd(a, b) = ax + by} The
coefficients \eqn{x} and \eqn{y} are known as Bezout's coefficients and can
be zero or negative.
}
\examples{
gcd_extended(99, 78)
gcd_extended(55, 45)
}
\references{
Bezout's identity. (2017, May 12). In Wikipedia, The Free
Encyclopedia. From
https://en.wikipedia.org/w/index.php?title=B%C3%A9zout%27s_identity&oldid=780050687
Cormen, T., Leiserson, C., Rivest, R., & Stein, C. (2009). Introduction to
algorithms (3rd ed., pp. 937-938). Cambridge (Inglaterra): Mit Press.
}
