% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gcd_subtraction}
\alias{gcd_subtraction}
\title{Computes the Greatest Common Divisor of two integers using the Euclidean subtraction algorithm.}
\usage{
gcd_subtraction(a, b)
}
\arguments{
\item{a}{First integer}

\item{b}{Second integer}
}
\value{
The greatest common divisor
}
\description{
Computes the greatest common divisor using the original implementation of the algorithm by
Euclid. In this version of the algorithm, the quotient \eqn{q_k} and remainder \eqn{r_k}
is calculated by repeated division by subtraction. Therefore, in this implementation of the
greatest common divisor algorithm, the modulo operation in the division implementation is
replaced by division by repeated subtraction.
}
