% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integers.R
\name{is.coprime}
\alias{is.coprime}
\title{Tests if two integers are coprime.}
\usage{
is.coprime(a, b)
}
\arguments{
\item{a}{First integer}

\item{b}{Second integer}
}
\value{
TRUE if given integers are coprime, FALSE otherwise
}
\description{
Two integers \eqn{a} and \eqn{b} are said to be relatively prime (also called
relatively prime) if they share no positive divisors except 1.
}
\examples{
is.coprime(2, 4)
is.coprime(2, 3)
}
\references{
Cormen, T., Leiserson, C., Rivest, R., & Stein, C. (2009).
Introduction to algorithms (3rd ed., pp. 931). Cambridge (Inglaterra): Mit
Press. Weisstein, Eric W. "Relatively Prime." From MathWorld--A Wolfram Web
Resource. http:#mathworld.wolfram.com/RelativelyPrime.html
}
