% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integers.R
\name{is.even}
\alias{is.even}
\title{Tests if a given value \eqn{n} is even or not.}
\usage{
is.even(n)
}
\arguments{
\item{n}{Value to test}
}
\value{
TRUE if n is even, FALSE otherwise
}
\description{
An even number is defined as an integer with the form \eqn{n = 2k} where
\eqn{k} is also an integer. Put differently, an even number is not odd and is
thus evenly divisible by 2.
}
\examples{
is.even(5)
is.even(4)
is.even(4.23)
}
\references{
Weisstein, Eric W. "Even Number." From MathWorld--A Wolfram Web
Resource. http:#mathworld.wolfram.com/EvenNumber.html
}
