% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integers.R
\name{is.odd}
\alias{is.odd}
\title{Tests whether a given given value n is odd.}
\usage{
is.odd(n)
}
\arguments{
\item{n}{Value to test}
}
\value{
TRUE if n is odd, FALSE otherwise
}
\description{
An odd number is an integer that has the form \eqn{n = 2k + 1} for an integer
\eqn{k}. In other words, an odd number is an integer that is not evenly
divisible by 2.
}
\examples{
is.odd(5)
is.odd(4)
is.odd(5.25)
}
\references{
Weisstein, Eric W. "Odd Number." From MathWorld--A Wolfram Web
  Resource. http:#mathworld.wolfram.com/OddNumber.html
}
