% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primalitytests.R
\name{is.prime}
\alias{is.prime}
\title{Tests whether a given value n is prime}
\usage{
is.prime(n)
}
\arguments{
\item{n}{Value to test}
}
\value{
TRUE if n is prime, FALSE otherwise
}
\description{
A prime number is defined as a positive integer, \eqn{n > 1} that has no
positive divisors other than 1 and itself.
}
\examples{
is.prime(9)
is.prime(3)
is.prime(10927)
}
\references{
Weisstein, Eric W. "Prime Number." From MathWorld--A Wolfram Web
Resource. http:#mathworld.wolfram.com/PrimeNumber.html
}
