% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integers.R
\name{is.square}
\alias{is.square}
\title{Tests if a given integer is a square number.}
\usage{
is.square(n)
}
\arguments{
\item{n}{Value to test}
}
\value{
TRUE if value is a square number, FALSE otherwise
}
\description{
A square number is an integer that is the square of an integer.
}
\examples{
is.square(25)
is.square(9)
is.square(3)
}
\references{
Barnes, C. (2004). Integer Factorization Algorithms (1st ed.).
Corvallis, OR: Department of Physics, Oregon State University.
}
