% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{isodd}
\alias{isodd}
\title{Tests whether a given given value n is odd.}
\usage{
isodd(n)
}
\arguments{
\item{n}{Value to test}
}
\value{
TRUE if n is odd, FALSE otherwise
}
\description{
An odd number is an integer that has the form \eqn{n = 2k + 1} for an integer
\eqn{k}. In other words, an odd number is an integer that is not evenly
divisible by 2.
}
\examples{
isodd(3)
isodd(4)
isodd(3.2)
}
\references{
Weisstein, Eric W. "Odd Number." From MathWorld--A Wolfram Web
  Resource. http://mathworld.wolfram.com/OddNumber.html
}
