% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{issquare}
\alias{issquare}
\title{Tests if a given integer is a square number.}
\usage{
issquare(n)
}
\arguments{
\item{n}{Value to test}
}
\value{
TRUE if value is a square number, FALSE otherwise
}
\description{
A square number is an integer that is the square of an integer.
}
\references{
Barnes, C. (2004). Integer Factorization Algorithms (1st ed.).
Corvallis, OR: Department of Physics, Oregon State University.
}
