% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lcm}
\alias{lcm}
\title{Computes the least common multiple of two integers, a and b.}
\usage{
lcm(a, b)
}
\arguments{
\item{a}{First integer}

\item{b}{Second integer}
}
\value{
the least common multiple of the two given integers
}
\description{
The least common multiple is defined as the smallest possible positive integer that 
is divisible by both \eqn{a} and \eqn{b}, and is usually denoted \eqn{lcm(a,b)} or 
\eqn{LCM(a,b)}. The function implementation for finding the least common multiple takes advantage 
of the relation between the greatest common divisor and the least common multiple, such that:
\deqn{lcm(a,b) = \bigg(\frac{|a|}{gcd(a,b)} \bigg) \space |b|}
}
\examples{
lcm(21, 6)
lcm(15, 3)
}
