% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{miller_rabin}
\alias{miller_rabin}
\title{Tests an integer's primality using the Miller-Rabin test.}
\usage{
miller_rabin(n, k)
}
\description{
Tests an integer's primality using the Miller-Rabin test.
}
