% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mod_expo}
\alias{mod_expo}
\title{Computes large modular exponents with modular exponentiation.}
\usage{
mod_expo(a, n, m)
}
\arguments{
\item{a}{the base integer, as in a^n mod m}

\item{n}{the exponent power}

\item{m}{the modulus}
}
\value{
the computed exponent mod m
}
\description{
Computes large modular exponents with modular exponentiation.
}
