% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor.R
\name{pollardrho}
\alias{pollardrho}
\title{Implementation of Pollard's rho algorithm for factorizing an
integer \eqn{n} into two non-trivial prime numbers.}
\usage{
pollardrho(n)
}
\arguments{
\item{n}{Integer to be factored into product of smaller integers.}
}
\value{
Vector containing the two prime factors of \eqn{n}, should they exist.
}
\description{
Implementation of Pollard's rho algorithm for factorizing an
integer \eqn{n} into two non-trivial prime numbers.
}
\examples{
pollardrho(9)
pollardrho(18364)
pollardrho(13)
}
\references{
Barnes, C. (2004). Integer Factorization Algorithms (1st ed.).
  Corvallis, OR: Department of Physics, Oregon State University. Fermat's
  factorization method. (2017, January 31). In Wikipedia, The Free
  Encyclopedia. From
  https:#en.wikipedia.org/w/index.php?title=Fermat%27s_factorization_method&oldid=763010603
}
