% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorial.R
\name{rising.factorial}
\alias{rising.factorial}
\title{Computes the rising factorial. Also known as the Pochhammer symbol.}
\usage{
rising.factorial(x, n)
}
\arguments{
\item{x}{Integer. The value will be rounded down if the value is not an
integer.}

\item{n}{Integer. The value will be rounded down if the value is not an
integer.}
}
\value{
If x is an integer, the rising factorial for an integer :math:`n`,
  \eqn{(x)_{n}}. If x is a string, the output is the symbolic representation
  of the rising factorial.
}
\description{
The rising factorial, \eqn{x^{(n)}} (sometimes denoted \eqn{\langle x
\rangle_n}) is also known as the Pochhammer symbol in other areas of
mathematics.
}
\details{
The rising factorial is related to the gamma function \eqn{\Gamma (z)}.
\deqn{x^{(n)} \equiv \frac{\Gamma (x + n)}{\Gamma (n)}} where \eqn{x^(0) =
1}. The rising factorial is related to the falling factorial by:
\deqn{x^{(n)} = (-x)_n (-1)^n}
}
\examples{
risingfactorial(10, 6)
risingfactorial('x', 4)
}
\references{
Falling and rising factorials. (2017, June 8). In Wikipedia, The
Free Encyclopedia. From
https://en.wikipedia.org/w/index.php?title=Falling_and_rising_factorials&oldid=784512036

Weisstein, Eric W. "Rising Factorial." From MathWorld--A Wolfram Web
Resource. http://mathworld.wolfram.com/RisingFactorial.html
}
