% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorial.R
\name{stirling}
\alias{stirling}
\title{Approximates a factorial of an integer \eqn{n} using Stirling's
Approximation. Specifically, the approximation is done using a method
developed by Gosper.}
\usage{
stirling(n)
}
\arguments{
\item{n}{desired integer to approximate factorial}
}
\description{
Stirling's approximation is a method of approximating a factorial \eqn{n!}.
As the value of \eqn{n} increases, the more exact the approximation becomes;
however, it still yields almost exact results for small values of \eqn{n}.
The approximation used is given by Gosper, which is noted to be a better
approximation to \eqn{n!} and also results in a very close approximation to
\eqn{0! = 1}. \deqn{n! \approx \sqrt{(2n + \frac{1}{3})\pi} n^n e^{-n}}
}
\examples{
stirling(0)
stirling(5)
stirling(50)
}
\references{
Stirling's approximation. (2017, March 8). In Wikipedia, The
  Free Encyclopedia. From
  https:#en.wikipedia.org/w/index.php?title=Stirling%27s_approximation&oldid=769328178
   Weisstein, Eric W. "Stirling's Approximation." From MathWorld--A Wolfram
  Web Resource. http:#mathworld.wolfram.com/StirlingsApproximation.html
}
