% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{three_n_one}
\alias{three_n_one}
\title{Computes the famous \eqn{3n + 1} sequence, also known as the Collatz
conjecture.}
\usage{
three_n_one(n)
}
\arguments{
\item{n}{Starting integer to begin \eqn{3n + 1} process}
}
\value{
vector containing sequence generated by the Collatz conjecture. If n
  is 1, 1 is returned.
}
\description{
The Collatz conjecture, also known as the \eqn{3n + 1} problem, is a
currently unsolved problem in number theory that is stated as: Start with a
positive integer \eqn{n} If \eqn{n} is even, divide \eqn{n} by 2 If \eqn{n}
is odd, multiply by 3 and add 1 to obtain \eqn{3n + 1} Repeat this process
until the sequence reaches 1. It is unknown if there is a starting integer
\eqn{n} that does not result in the sequence ending with 1.
}
\references{
3x + 1 problem. (2017, May 13). In Wikipedia, The Free
Encyclopedia. from
https://en.wikipedia.org/w/index.php?title=3x_%2B_1_problem&oldid=780191927
Moler, C. (2011). Numerical computing with MATLAB (1st ed.). Philadelphia,
Pa: Society for Industrial & Applied Mathematics.
}
