% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{binomial_coefficient}
\alias{binomial_coefficient}
\title{Calculates the binomial coefficient with several different algorithm methods
available.}
\usage{
binomial_coefficient(n, k, method = "multiplicative")
}
\arguments{
\item{n}{Number of possibilities}

\item{k}{number of unordered outcomes}

\item{method}{selects the algorithm to use for calculating the binomial
coefficient. Options include the multiplicative (default), recursive, or
factorial methods. Please note the recursive method can be very slow and
is not recommended over other algorithms.}
}
\value{
The binomial coefficient
}
\description{
The binomial coefficient equation (in compact form) is defined as:
\deqn{\binom{n}{k} = \frac{n!}{k!(n-k)!} \qquad 0 \leq k \leq n} The
multiplicative method for computing the binomial coefficient is more
efficient than the compact form calculation and is defined as:
\deqn{\binom{n}{k} = \frac{n(n-1)(n-2) \cdots (n-(k-1))}{k(k-1)(k-2) \cdots
1} = \prod^k_{i=1} \frac{n + 1 - i}{i}} The recursive method of the binomial
coefficient calculation is defined as: \deqn{\binom{n}{k} = \binom{n - 1}{n
- k} + \binom{n - 1}{k} \qquad for n, k: 1 \leq k \leq n - 1} With boundary
values: \deqn{\binom{n}{0} = \binom{n}{n} = 1}
}
\examples{
binomial_coefficient(5, 3)
binomial_coefficient(6, 4, 'factorial')
}
\references{
Binomial coefficient. (2017, April 17). In Wikipedia, The Free
  Encyclopedia. From
  https:#en.wikipedia.org/w/index.php?title=Binomial_coefficient&oldid=775905810
   Binomial coefficients. Encyclopedia of Mathematics. From
  http:#www.encyclopediaofmath.org/index.php?title=Binomial_coefficients&oldid=39155
   Weisstein, Eric W. "Binomial Coefficient." From MathWorld--A Wolfram Web
  Resource. http:#mathworld.wolfram.com/BinomialCoefficient.html
}
