% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{binomial_multiplicative}
\alias{binomial_multiplicative}
\title{Calculates the binomial coefficient using the multiplicative equation.}
\usage{
binomial_multiplicative(n, k)
}
\arguments{
\item{n}{Number of possibilities}

\item{k}{number of unordered outcomes}
}
\value{
The binomial coefficient
}
\description{
The multiplicative method for computing the binomial coefficient is more
efficient than the compact form calculation and is defined as:
\deqn{\binom{n}{k} = \frac{n(n-1)(n-2) \cdots (n-(k-1))}{k(k-1)(k-2) \cdots
1} = \prod^k_{i=1} \frac{n + 1 - i}{i}}
}
\references{
Binomial coefficient. (2017, April 17). In Wikipedia, The Free
  Encyclopedia. From
  https://en.wikipedia.org/w/index.php?title=Binomial_coefficient&oldid=775905810
   Binomial coefficients. Encyclopedia of Mathematics. From
  http://www.encyclopediaofmath.org/index.php?title=Binomial_coefficients&oldid=39155
   Weisstein, Eric W. "Binomial Coefficient." From MathWorld--A Wolfram Web
  Resource. http://mathworld.wolfram.com/BinomialCoefficient.html
}
