% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorial.R
\name{factorial}
\alias{factorial}
\title{Calculate the factorial \eqn{n!} of an integer \eqn{n}}
\usage{
factorial(n)
}
\arguments{
\item{n}{Value to calculate}
}
\value{
the factorial of the integer
}
\description{
Factorials are denoted for a positive integer \eqn{x} as \eqn{x!} and are
defined as: \deqn{n! = (n)(n - 1)(n - 2) \cdots (2)(1)} For example, the
factorial of 5 is written as: \deqn{5! = (5)(4)(3)(2)(1) = 120}
}
\references{
Press, W., Teukolsky, S., Vetterling, W., & Flannery, B. (2007).
  Numerical recipes (3rd ed.). Cambridge: Cambridge University Press.
  Weisstein, Eric W. "Factorial." From MathWorld--A Wolfram Web Resource.
  http://mathworld.wolfram.com/Factorial.html
}
