% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcd.R
\name{gcd}
\alias{gcd}
\title{Calculates the Greatest Common Divisor.}
\usage{
gcd(a, b, method = c("division", "subtraction", "recursive"))
}
\arguments{
\item{a}{First integer}

\item{b}{Second integer}

\item{method}{selects the algorithm used to compute the GCD. Options are recursive (default), subtraction, and division.}
}
\value{
The greatest common divisor
}
\description{
The Euclidean division implementation of the greatest common divisor
algorithm computes a quotient \eqn{q_k} and a remainder \eqn{r_k} at each
step \eqn{k} from the two numbers \eqn{a} and \eqn{b}. The quotient component
of Euclidean division is not used in the algorithm, thus only modulo
operations are required as that only returns the remainder.
}
\details{
The subtraction method for computing the GCD is the original implementation of the algorithm by
Euclid. In this version of the algorithm, the quotient \eqn{q_k} and remainder \eqn{r_k}
is calculated by repeated division by subtraction. Therefore, in this implementation of the
greatest common divisor algorithm, the modulo operation in the division implementation is
replaced by division by repeated subtraction.
}
\examples{
gcd(24, 30)
gcd(9, 5)
gcd(193752, 234574)
}
\references{
Cormen, T., Leiserson, C., Rivest, R., & Stein, C. (2009).
  Introduction to algorithms (3rd ed., pp. 928-930). Cambridge (Inglaterra):
  Mit Press. Euclidean algorithm. (2017, May 18). In Wikipedia, The Free
  Encyclopedia. From
  https:#en.wikipedia.org/w/index.php?title=Euclidean_algorithm&oldid=780973502
   Euclidean division. (2017, May 10). In Wikipedia, The Free Encyclopedia.
  From
  https:#en.wikipedia.org/w/index.php?title=Euclidean_division&oldid=779699188
}
