% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcd.R
\name{gcd}
\alias{gcd}
\title{Calculates the greatest common divisor of two integers using the Euclidean
algorithm.}
\usage{
gcd(a, b, method = c("recursive", "division", "subtraction"))
}
\arguments{
\item{a}{First integer}

\item{b}{Second integer}

\item{method}{Optional, select which algorithm to use when calculating GCD}
}
\value{
the greatest common divisor
}
\description{
The greatest common divisor of two integers \eqn{a} and \eqn{b} is the
greatest of the common divisors between two integers, for example
\eqn{gcd(32, 24) = 8}. Several implementations of the Euclidean algorithm are
available including a recursive method, division and subtraction (the
original version of the algorithm).
}
\examples{
gcd(24, 32)
gcd(125, 25, 'division')
gcd(9, 5)
gcd(9, 5, 'subtraction')
}
