% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{isprime}
\alias{isprime}
\title{Tests whether a given value n is prime with a naive test.}
\usage{
isprime(n)
}
\arguments{
\item{n}{Value to test}
}
\value{
TRUE if n is prime, FALSE otherwise
}
\description{
A prime number is defined as a positive integer, \eqn{n > 1} that has no
positive divisors other than 1 and itself.
}
\references{
Weisstein, Eric W. "Prime Number." From MathWorld--A Wolfram Web
Resource. http://mathworld.wolfram.com/PrimeNumber.html
}
