% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lucas_lehmer}
\alias{lucas_lehmer}
\title{Performs the Lucas-Lehmer primality test for determining if a Mersenne
number is prime.}
\usage{
lucas_lehmer(p)
}
\arguments{
\item{p}{integer}
}
\value{
TRUE if integer is prime (and a Mersenne Number), FALSE otherwise
}
\description{
Performs the Lucas-Lehmer primality test for determining if a Mersenne
number is prime.
}
\references{
Weisstein, Eric W. "Lucas-Lehmer Test." From MathWorld--A
  Wolfram Web Resource. 
  http://mathworld.wolfram.com/Lucas-LehmerTest.html
  
  Weisstein, Eric W. "Mersenne Number." From MathWorld--A Wolfram Web
  Resource. 
  http://mathworld.wolfram.com/MersenneNumber.html
  
  Weisstein, Eric W. "Primality Test." From MathWorld--A Wolfram Web Resource.
  http://mathworld.wolfram.com/PrimalityTest.html
}
