% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorial.R
\name{ramanujan}
\alias{ramanujan}
\title{Approximates the factorial :math:`n!` given an integer :math:`n` using
Ramanujan's formula. Ramanujan's formula is just as or more accurate than
several other factorial approximation formulas.}
\usage{
ramanujan(n)
}
\arguments{
\item{n}{integer to approximate factorial}
}
\description{
Ramanujan's formula is another factorial approximation method known for its
accuracy in comparison to other factorial approximation approaches including
Stirling's and Gosper's approximations. Ramanujan's formula is defined as:
\deqn{n! \approx \sqrt{\pi} \left(\frac{n}{e}\right)^n \sqrt[6]{8n^3 + 4n^2
+ n + \frac{1}{30}}}
}
\examples{
ramanujan(10)
ramanujan(1)
ramanujan(5)
}
\references{
Mortici, Cristinel. On Gosper's Formula for the Gamma Function.
  Valahia University of Targoviste, Department of Mathematics. Retrieved
  from http:#files.ele-math.com/articles/jmi-05-53.pdf
}
