% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorial.R
\name{stirlingln}
\alias{stirlingln}
\title{Approximates the factorial of n using the approximation given by Ramanujan
in his lost notebook (Ramanujan 1988, as cited in Wikipedia). Computing the
factorial in logarithmic form is useful as it helps avoid overflow when n is
large. As values of n increase, the approximation given becomes more exact.}
\usage{
stirlingln(n)
}
\arguments{
\item{n}{desired integer to approximate factorial}
}
\description{
It is often useful to compute the logarithmic form of the factorial and
convert it to exponent form to avoid overflow. The approximation is an
alternative approach given by Srinivasa Ramanujan (Ramanujan 1988). \deqn{ln
n! \approx n ln n - n + \frac{1}{6} ln(n(1 + 4n(1 + 2n))) + \frac{1}{2} ln
\pi}
}
\examples{
stirlingln(50)
stirlingln(10)
}
\references{
Stirling's approximation. (2017, March 8). In Wikipedia, The
  Free Encyclopedia. From
  https:#en.wikipedia.org/w/index.php?title=Stirling%27s_approximation&oldid=769328178
}
